Move DTRSMinimizer into TrustRegionMinimizer re #17645
