+++ Framework/CurveFitting/inc/MantidCurveFitting/FuncMinimizers/TrustRegionMinimizer.h
@@ -48,6 +48,8 @@ public:
   bool iterate(size_t) override;
   /// Return current value of the cost function
   double costFunctionVal() override;
+  /// Name of the minimizer.
+  std::string name() const override;
 
 private:
   /// Evaluate the fitting function and calculate the residuals.
@@ -58,12 +60,12 @@ private:
   void evalHF(const DoubleFortranVector &x, const DoubleFortranVector &f,
               DoubleFortranMatrix &h) const;
   /// Find a correction vector to the parameters.
-  virtual void calculateStep(const DoubleFortranMatrix &J,
+  void calculateStep(const DoubleFortranMatrix &J,
                              const DoubleFortranVector &f,
                              const DoubleFortranMatrix &hf,
                              const DoubleFortranVector &g, double Delta,
                              DoubleFortranVector &d, double &normd,
-                             const NLLS::nlls_options &options) = 0;
+                             const NLLS::nlls_options &options);
 
   /// Stored cost function
   boost::shared_ptr<CostFunctions::CostFuncLeastSquares> m_leastSquares;
@@ -81,6 +83,17 @@ private:
   NLLS::nlls_inform m_inform;
   /// Temporary and helper objects
   NLLS::NLLS_workspace m_workspace;
+
+  void solveStep(const DoubleFortranMatrix &J, const DoubleFortranVector &f,
+    const DoubleFortranMatrix &hf, double Delta,
+    DoubleFortranVector &d, double &normd,
+    const NLLS::nlls_options &options);
+
+  // Used for calculating step
+  DoubleFortranMatrix m_A, m_ev;
+  DoubleFortranVector m_ew, m_v, m_v_trans, m_d_trans;
+  NLLS::all_eig_symm_work m_all_eig_symm_ws;
+  DoubleFortranVector m_scale;
 };
 
 } // namespace FuncMinimisers
